requirements.txt
asgiref==3.10.0
certifi==2025.10.5
charset-normalizer==3.4.4
Django==5.2.7
django-environ==0.12.0
idna==3.11
marshmallow==3.26.1
packaging==25.0
psycopg2-binary==2.9.11
requests==2.32.5
sqlparse==0.5.3
transbank-sdk==6.1.0
tzdata==2025.2
urllib3==2.5.0
gunicorn==21.2.0
whitenoise==6.7.0
